ensure server handshake nextStatus and protocolVersion cannot be changed through public apis
