+++ api/src/main/java/com/velocitypowered/api/event/connection/ServerHandshakeEvent.java
@@ -19,6 +19,7 @@ package com.velocitypowered.api.event.connection;
 
 import com.google.common.base.Preconditions;
 import com.velocitypowered.api.event.annotation.AwaitingEvent;
+import com.velocitypowered.api.network.ProtocolVersion;
 import com.velocitypowered.api.proxy.Player;
 import com.velocitypowered.api.proxy.player.ServerHandshake;
 import com.velocitypowered.api.proxy.server.RegisteredServer;
@@ -38,24 +39,32 @@ import org.checkerframework.checker.nullness.qual.Nullable;
 public final class ServerHandshakeEvent {
   private final Player player;
   private final RegisteredServer server;
+  private final ProtocolVersion protocolVersion;
+  private final int nextStatus;
   private final ServerHandshake originalHandshake;
   private @Nullable ServerHandshake handshake;
 
   /**
    * Creates a new instance.
    *
-   * @param player the player connecting to the server
-   * @param server the server the player is connecting to
+   * @param player            the player connecting to the server
+   * @param server            the server the player is connecting to
+   * @param protocolVersion   the protocol version to send to the server
+   * @param nextStatus        the next status to send to the server
    * @param originalHandshake the original {@link ServerHandshake} to send to the server
    */
-  public ServerHandshakeEvent(Player player, RegisteredServer server, ServerHandshake originalHandshake) {
-    this.player = player;
-    this.server = server;
+  public ServerHandshakeEvent(Player player, RegisteredServer server,
+      ProtocolVersion protocolVersion, int nextStatus, ServerHandshake originalHandshake) {
+    this.player = Preconditions.checkNotNull(player, "player");
+    this.server = Preconditions.checkNotNull(server, "server");
+    this.protocolVersion = Preconditions.checkNotNull(protocolVersion, "protocolVersion");
+    this.nextStatus = nextStatus;
     this.originalHandshake = Preconditions.checkNotNull(originalHandshake);
   }
 
   /**
    * Returns the player connecting to the server.
+   *
    * @return the player connecting to the server
    */
   public Player getPlayer() {
@@ -64,10 +73,29 @@ public final class ServerHandshakeEvent {
 
   /**
    * Returns the server the player is connecting to.
+   *
    * @return the server the player is connecting to
    */
   public RegisteredServer getServer() {
     return server;
+  }
+
+  /**
+   * Returns the protocol version to send to the server.
+   *
+   * @return the protocol version
+   */
+  public ProtocolVersion getProtocolVersion() {
+    return protocolVersion;
+  }
+
+  /**
+   * Returns the next status to send to the server.
+   *
+   * @return the next status
+   */
+  public int getNextStatus() {
+    return nextStatus;
   }
 
   public ServerHandshake getOriginalHandshake() {
